pyproject.toml
[project]
name = "sae-genomics"
version = "0.1.0"
description = "Sparse Autoencoder training for tahoe x1 single-cell foundation model"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["sparse-autoencoders", "single-cell", "genomics", "interpretability", "tahoe-x1"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    # Core ML/DL
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "einops>=0.7.0",

    # SAE training (SAELens will be installed from submodule)
    "sae-lens>=6.20.0",

    # Scientific computing
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.1.0",

    # Statistics
    "statsmodels>=0.14.0",

    # API clients & data fetching
    "requests>=2.31.0",
    "aiohttp>=3.9.0",

    # Bioinformatics
    "mygene>=3.2.0",
    "pronto>=2.5.0",  # For ontology parsing (HPO)

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # Experiment tracking
    "wandb>=0.16.0",

    # Progress bars
    "tqdm>=4.66.0",

    # Utilities
    "rich>=13.7.0",  # Beautiful terminal output
    "typer>=0.9.0",  # CLI framework
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",

    # Code quality
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Type stubs
    "pandas-stubs>=2.1.0",
    "types-requests>=2.31.0",
    "types-pyyaml>=6.0.0",
]

notebooks = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.27.0",
    "ipywidgets>=8.1.0",
    "nb-black>=1.0.7",  # Black formatter for notebooks
]

validation-extra = [
    # Additional validation tools
    "gseapy>=1.1.0",  # Gene set enrichment analysis
    "gprofiler-official>=1.0.0",  # GO enrichment
    "bioservices>=1.11.0",  # Access to biological databases
]

all = [
    "sae-genomics[dev,notebooks,validation-extra]",
]

[project.urls]
Homepage = "https://github.com/yourusername/sae-genomics"
Documentation = "https://github.com/yourusername/sae-genomics/blob/main/README.md"
Repository = "https://github.com/yourusername/sae-genomics"
Issues = "https://github.com/yourusername/sae-genomics/issues"

[project.scripts]
sae-train = "sae_genomics.cli:train"
sae-validate = "sae_genomics.cli:validate"
sae-export = "sae_genomics.cli:export_results"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sae_genomics"]

[tool.uv]
dev-dependencies = [
    "sae-genomics[dev]",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | external
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "external",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["sae_genomics"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
exclude = [
    "external/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "sae_lens.*",
    "mygene.*",
    "pronto.*",
    "gseapy.*",
    "gprofiler.*",
    "bioservices.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/sae_genomics",
    "--cov-report=term-missing",
    "--cov-report=html",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/sae_genomics"]
omit = [
    "*/tests/*",
    "*/external/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
